requirements.txt
selenium==4.16.0
python-dotenv==1.0.0
httpx==0.26.0